Add a mechanism to substitute inferred dependencies with alternatives (per-groupId).  This allows for automatically generated dependencies (say dagger_compiler) to have their dependency on :dagger rewritten wtihin the package, for cases where you alter the target name via build_substitutions.
This substitution is done per-group, as it's really targetted at internal inter-depenencies of complex packages, not a general sweeping replacement mechanism.  The latter is dangerous, because mostly you want to rewrite some package-local relationships, but leave the rest of the world looking at the public target as they normally would.  This feature seeks to strike the balance of eliminating a lot of need for extra build_substitutes, without making hard to specify and hard to implement targeted search and replace.