Support configurable artifact properties.
Rework the API for maven_repository_specification to allow each artifact to have a property dictionary attached.  The first two properties supported are sha256 (and this replaces just passing a string as the sha) and insecure, which replaces the insecure_artifacts approach, by making you just add a property to it if you don't include the sha.  This also allows for later expansion of the feature set for per-artifact manipulation.  This could replace build_substitutions also, but that will happen in a later commit.

Also fixes the test rig a bit, to make it so that a missing declaration of a suite will be caught.  Minor validation, but helpful.

Lastly, support the older configuration style, but print deprecation messages, and in the process, rework validation of the input and print decent errors.