For a given pom, get the 'effective pom' that incorporates all the metadata from the parent.
Note, this will be a limited form of this, only incorporating the core metadata, the dependencies, the properties, and the dependencyManagement sections. The lack of recursion makes it more awkward to write, so we'll go for the three top level nodes we need.

This commit does not obtain the implicit dependency information from the other two sides, just makes the metadata complete, so that such inferred metadata can be valid.