Add in both some kotlin production and test files, but also an AutoValue to demonstrate multiple annotation processors working, hooked up by the maven integration, with kotlin.
In particular, also use the reified inline overload of assertFailsWith from kotlin.test to force the use of kotlin inlining (reproducing a bug whereby ijar strips out function bodies even when necessary to the compiler for inlining).  This should, as of this commit, cause the tests to not merely fail, but for the compiler to crash.