Add a case to the test_workspace which proves that the compile-time transitive deps are present.
If B deps on A as a part of API, e.g. inheriting from a class, A must be in the compilation classpath.  This case adds a java_library example and a kt_jvm_library example of just such a case (rxjava2's Flowable and reactive-streaming's Publisher).  This succeeds in the java_library case, so it's clear bazel_maven_repository's construction of a JavaInfo is sound. It fails in kotlin (even java compiled by the kotlin rules' kotlinc invocation) because it has half-baked strict-deps support, and ends up not consuming the proper classpath. Because it fails, it's commented out with a note pointing at the issue bazelbuild/rules_kotlin#132.  This can be commented out at hte point where it's fixed.