Fix #62 by fixing boolean coercion.
Turns out bool(some_string) does odd things, including returning true on non-empty strings.  Should not have been using magic coercion, but instead needed to test that the content of the string was what Maven interpreted as 'true' and in all other cases return false.

So do that now. Also add a test, and add an example of that with an explanation in the test_workspace.  (can't really test the failure because we can't fail the build-under-test).