Infer the packaging type from the pom.
Moves pom fetching into independent workspaces, earlier, and discovers the pom packaging type from the pom.xml as-parsed, avoiding the need to specify it in the artifact dictionary (unless a classifier is being used, in which case it must precede the classifier, as that's the standard format of the maven slug).

Also factor out an faux-enum.  Also add a bunch of tests.  Also move some things into globals.bzl which require consistent discovery across the tool (like the workspace/path of a given pom file, so this can be turned into a real path and read where needed, by different rules).