Relax some of hte name mangling that bazel_maven_repository does, so that artifact_ids that don't have `.` in their names aren't mangled.
In particular, `-` is now not mangled, as a lot of artifact_ids have this, but bazel is perfectly capable of handling this when it comes to targets.

Also add a legacy mode, so that projects upgrading can just flag on to have compatibility aliases generated with the old names, while they adjust to the new naming implications.

The upshot is that now `com.google.dagger:dagger-compiler:2.24` would be referenced as `@maven//com/google/dagger:dagger-compiler`