Move missing-deps error reporting to the end of the loop.
This doesnt' collect ALL of the missing deps, because there may be missing transitive deps, but this moves at least all missing direct deps for all specified artifacts into one report, reducing the number of iterations. We still need to collect the un-specified transitive deps, but this is more invasive, and requires fetching more poms.  Doable, but in a different commit.