Initial version of the 2.0 line, which uses a delegated jvm binary to do the key work, using maven-archeologist as the resolution/fetching system.
While this is still mostly compatible with previous versions, some migrations will be needed (migrating to 1.2.0 is a reasonable half-step).

In particular, "-" is now permitted in artifact_id-based target names, and several legacy handling has been deleted (separate list of insecure artifacts, etc).

Most of what changes here is a complete gutting of the internals in favor of "kramer" (named for Carol Kramer, a pioneer in re-thinking archeology). Kramer uses maven-archeologist to do fast parallel fetching of metadata, model processing, and ends up replacing all the starlark based xml parser and pom handling infrastructure. It's fast. Really fast.

This will form the base for further feature work. There are definitely a few things TBD, including a bit better unit-testing (vs. integration testing which this leans on pretty heavily), as well as using maven-archeologist for the main artifact fetching (vs. just the metadata fetching), to benefit from shared caching infrastructure. There are lots of further features that are easier to make in this codebase, so we'll revisit a lot of prior requests in the 2.0 line.