Use kramer for the artifact fetch, directly.
This allows the artifact fetch/download to make use of the maven local repository cache in case (as is the case with .aar in all cases, and others on subsequent runs) the artifact has already been downloaded. It checks any sha256 hash given, but can download (and cache) insecurely.

Also rips out the extra starlark left over from the older process. What remains is only needed to support directly the needs of preparing the execution of kramer, or to support the legacy maven_jvm_artifact macro (used for build_snippets).

Also add integration tests for the process, ensuring .aars are unzipped appropriately, shas are compared, files are copied to the correct workspace location, etc.