Support source jar propagation.
Upgrades maven-archeologist to 0.0.4 which has sub-artifact fetchign capability.

Uses this to (if possible) download a sources.jar for the given artifact, and propagates it through a filegroup so raw_jvm_import appropriately specifies it. This permits IntelliJ (or anthing else that consumes source_jar metadata to provide the sources to the developer.